Split modules from highlight/engine
- Split functions from engine module into container-blocks and
  page-updates, making engine a type-only module
